[Team03][BE] 세번째 PR입니다 (#189)
* feat: spring-boot 설치

* feat: 도메인 제작 및 연관관계 수립

* feat: MemberRepository 제작

* feat: DB 설정

* feat: 도메인에 롬복 어노테이션 설정

* feat: service ,Repository ,Controller 추가

* feat: 잭슨 라이브러리에 사용하지 않는 프록시 조회는 놔두게 하기 위해 잭슨 의존성 추가

* feat: RoomQuantityDto 추가

* refactor: 병합 중 중복으로 올려진 gradle 설정 변경

* refactor: RoomRepository 에 JpaRepository 적용

* refactor: 도메인에 추가 내용 적용

* feat: Room 수량 관련 DTO 제작

* feat: Room 간단 정보 관련 DTO 제작

* refactor: 컨트롤러 이름에서 API 제외

* feat: Room 수량 및 간단 정보 관련 메소드 추가

* feat: Room 수량 및 간단 정보 관련 메소드 추가

* feat: entityManager 오류 해결 위해 javassist 의존성 추가

* feat: pull 병합

* refactor: 패키지 변경 작업

* refactor: RoomController url명 복수형으로 변경

* refactor: RequestRoomSearchInfo 를 RoomSearchInfoRequestDto 로 이름을 변경하여 네이밍 순서에 통일성 부여

* refactor: RoomController.loadInfoList() 의 매핑을 GetMapping으로 변경 하였습니다.

* refactor: RoomQuantityResponseDto 에서 문구 없이 데이터만 반환하도록 변경

* refactor: RoomSimpleInfoResponseDto 에서 정적 펙토리 메서드명을 매개변수 여러개 받는 것에 맞게 수정

* refactor: 데이터를 포맷해서 보내주는 내용들을 데이터만 보내주는 것으로 변경함

* refactor: 가격대별 방의 개수를 계산해서 보내주던 것을 가격 리스트만 보내는 것으로 변경

* refactor: 특정 변수타입 Long -> Integer 로 변경

+++ be/src/main/java/teamproject/airbnb/domain/Review.java
@@ -22,5 +22,5 @@ public class Review {
 	@JoinColumn(name = "ROOM_ID")
 	private Room room;
 
-	private Long Rating;
+	private Integer Rating;
 }
